adding  modified files  and deleted some files for code refactoring  [finishes #168360652]
