+++ server/configurations/db-config.js
@@ -1,0 +1,15 @@
+import { Pool } from 'pg';
+
+import dotenv from 'dotenv';
+
+
+dotenv.config();
+
+
+const connectionString = process.env.DATABASE_URL;
+
+const pool = new Pool({
+  connectionString,
+});
+
+export default pool;
