+++ server/routes/adminRoute.js
@@ -1,11 +1,11 @@
 
 import express  from 'express';
-import UserToMentor from '../controllers/userToMentor';
+import allMentors from '../controllers/mentors';
 import Authentication from '../middleware/Authentication';
 import adminCheck  from '../middleware/adminCheck';
 
 const route = express.Router();
 
-route.patch('/api/v1/user/:userId',[Authentication, adminCheck],UserToMentor);
+route.patch('/api/v1/user/:userId',[Authentication, adminCheck],allMentors.changeUserToMentor);
 
 module.exports = route;
