+++ server/routes/userRoute.js
@@ -1,18 +1,14 @@
 import express  from'express';
 import signinValidater  from'../middleware/signin';
-import signin  from'../controllers/userSignin';
-import signupValidater  from'../middleware/signup'; 
-import signup  from'../controllers/userSignup';
+import signing  from'../controllers/usersign';
+import signupValidater  from'../middleware/signup';
 import Authorisation  from'../middleware/Authentication';
-import allMentors from '../controllers/allMentors';
-import specificMentor from '../controllers/specificMentor';
-
-
+import allMentors from '../controllers/mentors';
 const route = express.Router();
 
-route.post('/api/v1/auth/signin',signinValidater,signin.signin);
-route.post('/api/v1/auth/signup',signupValidater,signup.signup);
-route.get('/api/v1/mentors',Authorisation,allMentors.MentorsDisplay);
-route.get('/api/v1/mentors/:mentorId',Authorisation, specificMentor.MentorDisplay);
+route.post('/api/v1/auth/signin',signinValidater,signing.signin);
+route.post('/api/v1/auth/signup',signupValidater,signing.signup);
+route.get('/api/v1/mentors',Authorisation,allMentors.mentorsDisplay);
+route.get('/api/v1/mentors/:mentorId',Authorisation,allMentors.specificMentorDisplay );
 
 export default route;
